// tests/test-duckdb/src/codegen/main.document_order-table.ts
/*
 File generated by Vexnor. Do not edit!
*/
import * as vexnor from "@vexnor/core";


export const DocumentOrder = vexnor.newSqlTable<{
   Select: IDocumentOrderSelect;
   Insert: IDocumentOrderInsert;
   Update: IDocumentOrderUpdate;
   Delete: true;
   Source: "@vexnor/test-duckdb:src/codegen";
}>( {
   crud: {
      select: true,
      insert: true,
      update: true,
      delete: true,
   },
   tableInfo: {
      name: "document_order",
      schema: "main",
   },
   pk: ["documentId"],
   dialect: "duckdb",
   source: "@vexnor/test-duckdb:src/codegen",
   columns: {

      /**
       * document_id undefined
       */
      documentId: "document_id",

      /**
       * account undefined
       */
      account: "account",

      /**
       * status undefined
       */
      status: "status",

      /**
       * shipping undefined
       */
      shipping: "shipping",

      /**
       * items undefined
       */
      items: "items",

      /**
       * tags undefined
       */
      tags: "tags",

      /**
       * created_at undefined
       */
      createdAt: "created_at",
   },
   jsonSchema: {
      createdAt: "Date",
   },
   dbSchema: {
      documentId: { dbType: "VARCHAR", type: vexnor.SqlLiteralType.String },
      account: { dbType: "STRUCT(account_id VARCHAR, email VARCHAR, status VARCHAR)", type: vexnor.SqlLiteralType.Json },
      status: { dbType: "VARCHAR", type: vexnor.SqlLiteralType.String },
      shipping: { dbType: "STRUCT(address STRUCT(street VARCHAR, city VARCHAR, country VARCHAR, geo STRUCT(latitude DOUBLE, longitude DOUBLE)), carrier STRUCT(\"name\" VARCHAR, tracking_id VARCHAR))", type: vexnor.SqlLiteralType.Json, nullable: true },
      items: { dbType: "STRUCT(product STRUCT(product_id VARCHAR, \"label\" VARCHAR, category VARCHAR), quantity INTEGER, unit_price DOUBLE, discounts STRUCT(code VARCHAR, amount DOUBLE)[])[]", type: vexnor.SqlLiteralType.Json },
      tags: { dbType: "VARCHAR[]", type: vexnor.SqlLiteralType.Json },
      createdAt: { dbType: "TIMESTAMP", type: vexnor.SqlLiteralType.Date },
   },
});
export type IDocumentOrderInsert = {
   documentId: string;
   account: unknown;
   status: string;
   shipping?: unknown | null;
   items: unknown;
   tags: unknown;
   createdAt: Date;
};

export type IDocumentOrderUpdate = Partial<IDocumentOrderInsert>;

export type IDocumentOrderSelect = {
   documentId: string;
   account: unknown;
   status: string;
   shipping: unknown | null;
   items: unknown;
   tags: unknown;
   createdAt: Date;
};

export type IDocumentOrderJson = vexnor.JsonRow<IDocumentOrderSelect>;